new commit

+++ labs/factorial.c
@@ -1,21 +1,18 @@
 #include<stdio.h>
-int factorial(int n)
+unsigned long int factorial(int n)
 {
-    int fact=1;
+    unsigned long int fact=1;
     for(int i=1;i<=n;i++)
     {
         fact=fact*i;
     }
     return fact;
 }
-
 int main()
 {
     int n;
-
     scanf("%d",&n);
-
-    printf("%d\n",factorial(n));
+    printf("%lu\n",factorial(n));
     return 0;
-    
+
 }
